+++ kaithem/src/version_info.py
@@ -1,2 +1,2 @@
-__version__ = "0.68.32 Production"
-__version_info__ = (0, 68, 32, "release", 0)
+__version__ = "0.68.33 Production"
+__version_info__ = (0, 68, 33, "release", 0)